+++ test/llm_composer/providers/utils_test.exs
@@ -1,0 +1,5 @@
+defmodule LlmComposer.Providers.UtilsTest do
+  use ExUnit.Case
+
+  alias LlmComposer.Providers.Utils
+end